requirements.txt
fastapi>=0.101.0
uvicorn>=0.23.2
tortoise-orm>=0.20.0
passlib>=1.7.4
bcrypt>=4.0.1
pyjwt>=2.3.0
python-multipart>=0.0.6
pandas>=2.1.1
dropbox>=11.36.2
openpyxl>=3.1.2